05 - data platform/my-taxi-pipeline/pipeline/assets/ingestion/requirements.txt
# Python dependencies for NYC taxi ingestion asset
pandas>=2.2.0
pyarrow>=14.0.0
python-dateutil>=2.8.0